done: secured edit, add, delete routes (fronted & backend). myspinnet page created. username on snippet showing. enhanced feature: search bar added.

+++ backend/controller/snippetController.js
@@ -3,7 +3,20 @@ const asyncHandler = require("express-async-handler")
 
 const getSnippets =async(req,res)=>{
     try{
-        const snips = await Snippet.find()
+        const search = req.query.search
+        let query = {};
+
+        if(search){
+            query = {
+                $or:[
+                    {title: {$regex: search, $options: "i"}},
+                    {description: {$regex: search, $options: "i"}},
+                    {tags: {$in: [new RegExp(search, "i")]}},
+                ]
+            }
+        }
+
+        const snips = await Snippet.find(query).populate("createdBy", "username _id").sort({createdAt: -1})
         res.json(snips)
     }catch(error){
         console.log(`Error connecting DB: ${error.message}`)
@@ -21,6 +34,16 @@ const getOneSnippet =asyncHandler(async(req,res)=>{
     res.status(200).json(snip)
 })
 
+const getMySnippets = asyncHandler(async(req,res)=>{
+    
+        const UserId = req.user.id; //verifyaccesstoken(in authmiddleware.js) provides this user
+        const snippets = await Snippet.find({createdBy: UserId}).sort({createdAt: -1})
+        
+        res.status(200).json(snippets)
+
+    
+})
+
 
 const createSnippet =asyncHandler(async(req,res)=>{
     const {title, description, tags, code}= req.body
@@ -28,12 +51,20 @@ const createSnippet =asyncHandler(async(req,res)=>{
         return res.status(400).json({message: "Title and Code fields are required."})
         // throw new Error("Title and code fields are required")
     }
-    const newSnippet = await Snippet.create({title, description, tags, code})
+    const newSnippet = await Snippet.create({title, description, tags, code,createdBy:req.user.id})
     res.status(201).json(newSnippet)
 })
 
 
 const editSnippet =asyncHandler(async(req,res)=>{
+    const snippet = await Snippet.findById(req.params.id);
+
+    if (!snippet) return res.status(404).json({ message: "Snippet not found" });
+
+    //check ownership
+    if (snippet.createdBy.toString() !== req.user.id) {
+      return res.status(403).json({ message: "Not authorized" });
+    }
     
     const updatedSnip = await Snippet.findByIdAndUpdate(req.params.id, req.body,{new:true} )
     if(!updatedSnip){
@@ -45,6 +76,15 @@ const editSnippet =asyncHandler(async(req,res)=>{
 
 
 const deleteSnippet =asyncHandler(async(req,res)=>{
+    const snippet = await Snippet.findById(req.params.id);
+
+    if (!snippet) return res.status(404).json({ message: "Snippet not found" });
+
+    // check ownership
+    if (snippet.createdBy.toString() !== req.user.id) {
+      return res.status(403).json({ message: "Not authorized" });
+    }
+
     let snip = await Snippet.findByIdAndDelete(req.params.id)
     if(!snip){
         return res.status(404).json({message: "Snippet not found."})
@@ -54,4 +94,4 @@ const deleteSnippet =asyncHandler(async(req,res)=>{
     res.status(200).json({message:"deleted snippet"})
 })
 
-module.exports = { getSnippets, getOneSnippet, createSnippet, editSnippet, deleteSnippet}
+module.exports = { getSnippets, getOneSnippet, createSnippet, editSnippet, deleteSnippet, getMySnippets}
